Added network interface monitoring script
If activated, checks if there has been any changes in the network. If
yes then restarts the ibp_server. Currently, in ibp_server 0.0.0.0
interface listening has few problems and we need to explicitely list all
the interface that we would like to listen to. Additionally, currently
ibp_server can not accomodate any interface changes which we would like
to fix in the future.